Add podman secrets, luks root encryption with TPM2
Using podman secrets to load secrets used in Pihole, v4l2rtspserver, and motion
Motion requires the secrets to be added to the config files before the motion container starts.
This is accomplished by using a bash container, and using a script to copy them into a temporary volume, replace the placeholder with each secret.
The volume is then deleted when the service is stopped.